chore: login refactoring
