+++ src/components/login/login.types.ts
@@ -19,6 +19,3 @@ import { SSOConfigDTO, SSOProvider } from '@Pages/GlobalConfigurations'
 export interface SSOConfigLoginList extends Pick<SSOConfigDTO, 'id' | 'active' | 'url'> {
     name: SSOProvider
 }
-export interface LoginCardProps {
-    renderContent: () => JSX.Element
-}
